# code/back-end/docs/swagger.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/api/feedback-cycles": {
      "get": {
        "operationId": "FeedbackCyclesController_findAll",
        "parameters": [
          {
            "name": "x-role",
            "in": "header",
            "description": "Caller role for RBAC",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "X-Role",
            "required": false,
            "schema": {
              "type": "string",
              "enum": [
                "superuser",
                "admin",
                "dean",
                "hod",
                "faculty",
                "student"
              ]
            }
          },
          {
            "in": "header",
            "name": "X-User-Id",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "X-User-Name",
            "required": false,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "summary": "List all feedback cycles",
        "tags": [
          "Feedback Cycles"
        ]
      },
      "post": {
        "operationId": "FeedbackCyclesController_create",
        "parameters": [
          {
            "name": "x-role",
            "in": "header",
            "description": "Caller role for RBAC",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "X-Role",
            "required": false,
            "schema": {
              "type": "string",
              "enum": [
                "superuser",
                "admin",
                "dean",
                "hod",
                "faculty",
                "student"
              ]
            }
          },
          {
            "in": "header",
            "name": "X-User-Id",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "X-User-Name",
            "required": false,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateFeedbackCycleDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "summary": "Create a feedback cycle (superuser/admin only)",
        "tags": [
          "Feedback Cycles"
        ]
      }
    },
    "/api/feedback-cycles/active": {
      "get": {
        "operationId": "FeedbackCyclesController_findActive",
        "parameters": [
          {
            "name": "x-role",
            "in": "header",
            "description": "Caller role for RBAC",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "X-Role",
            "required": false,
            "schema": {
              "type": "string",
              "enum": [
                "superuser",
                "admin",
                "dean",
                "hod",
                "faculty",
                "student"
              ]
            }
          },
          {
            "in": "header",
            "name": "X-User-Id",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "X-User-Name",
            "required": false,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "summary": "Get active feedback cycles",
        "tags": [
          "Feedback Cycles"
        ]
      }
    },
    "/api/feedback-cycles/state": {
      "get": {
        "operationId": "FeedbackCyclesController_getCycleState",
        "parameters": [
          {
            "name": "x-role",
            "in": "header",
            "description": "Caller role for RBAC",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "X-Role",
            "required": false,
            "schema": {
              "type": "string",
              "enum": [
                "superuser",
                "admin",
                "dean",
                "hod",
                "faculty",
                "student"
              ]
            }
          },
          {
            "in": "header",
            "name": "X-User-Id",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "X-User-Name",
            "required": false,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "summary": "Get current global cycle state (phase info)",
        "tags": [
          "Feedback Cycles"
        ]
      }
    },
    "/api/feedback-cycles/{id}": {
      "get": {
        "operationId": "FeedbackCyclesController_findOne",
        "parameters": [
          {
            "name": "x-role",
            "in": "header",
            "description": "Caller role for RBAC",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "X-Role",
            "required": false,
            "schema": {
              "type": "string",
              "enum": [
                "superuser",
                "admin",
                "dean",
                "hod",
                "faculty",
                "student"
              ]
            }
          },
          {
            "in": "header",
            "name": "X-User-Id",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "X-User-Name",
            "required": false,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "summary": "Get a single feedback cycle by ID",
        "tags": [
          "Feedback Cycles"
        ]
      },
      "patch": {
        "operationId": "FeedbackCyclesController_update",
        "parameters": [
          {
            "name": "x-role",
            "in": "header",
            "description": "Caller role for RBAC",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "X-Role",
            "required": false,
            "schema": {
              "type": "string",
              "enum": [
                "superuser",
                "admin",
                "dean",
                "hod",
                "faculty",
                "student"
              ]
            }
          },
          {
            "in": "header",
            "name": "X-User-Id",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "X-User-Name",
            "required": false,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateFeedbackCycleDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "summary": "Update a feedback cycle (superuser/admin only)",
        "tags": [
          "Feedback Cycles"
        ]
      },
      "delete": {
        "operationId": "FeedbackCyclesController_remove",
        "parameters": [
          {
            "name": "x-role",
            "in": "header",
            "description": "Caller role for RBAC",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "X-Role",
            "required": false,
            "schema": {
              "type": "string",
              "enum": [
                "superuser",
                "admin",
                "dean",
                "hod",
                "faculty",
                "student"
              ]
            }
          },
          {
            "in": "header",
            "name": "X-User-Id",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "X-User-Name",
            "required": false,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "summary": "Delete a feedback cycle (superuser/admin only)",
        "tags": [
          "Feedback Cycles"
        ]
      }
    },
    "/api/feedback-cycles/{id}/status": {
      "patch": {
        "operationId": "FeedbackCyclesController_updateStatus",
        "parameters": [
          {
            "name": "x-role",
            "in": "header",
            "description": "Caller role for RBAC",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "X-Role",
            "required": false,
            "schema": {
              "type": "string",
              "enum": [
                "superuser",
                "admin",
                "dean",
                "hod",
                "faculty",
                "student"
              ]
            }
          },
          {
            "in": "header",
            "name": "X-User-Id",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "X-User-Name",
            "required": false,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateCycleStatusDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "summary": "Update cycle status/phase (superuser/admin only)",
        "tags": [
          "Feedback Cycles"
        ]
      }
    },
    "/api/feedback-cycles/bulk": {
      "post": {
        "operationId": "FeedbackCyclesController_bulkImport",
        "parameters": [
          {
            "name": "x-role",
            "in": "header",
            "description": "Caller role for RBAC",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "X-Role",
            "required": false,
            "schema": {
              "type": "string",
              "enum": [
                "superuser",
                "admin",
                "dean",
                "hod",
                "faculty",
                "student"
              ]
            }
          },
          {
            "in": "header",
            "name": "X-User-Id",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "X-User-Name",
            "required": false,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/BulkImportCyclesDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "summary": "Bulk import historical feedback cycles (superuser/admin only)",
        "tags": [
          "Feedback Cycles"
        ]
      }
    },
...
  },
  "info": {
    "title": "ENDUR API",
    "description": "\n      REST API for the **ENDUR Performance Review & Feedback Management System**.\n\n      ## Role-Based Access Control\n      All protected endpoints require the `X-Role` header with one of:\n      `superuser` | `admin` | `dean` | `hod` | `faculty` | `student`\n\n      Additionally, pass `X-User-Id` and `X-User-Name` headers to attribute audit logs correctly.\n\n      ## Authentication\n      Use `POST /api/auth/login` to verify credentials. The response contains the user object\n      which you should store client-side (session) and use to populate role headers.\n      ",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "schemas": {
...
      "CreateFeedbackCycleDto": {
        "type": "object",
        "properties": {
          "cycleName": {
            "type": "string",
            "example": "Week 5 Formative Feedback"
          },
          "type": {
            "type": "string",
            "example": "weekly",
            "enum": [
              "weekly",
              "monthly",
              "midterm",
              "endterm"
            ]
          },
          "startTimestamp": {
            "type": "string",
            "example": "2026-04-20T00:00:00Z"
          },
          "endTimestamp": {
            "type": "string",
            "example": "2026-04-26T23:59:59Z"
          },
          "reflectionDeadline": {
            "type": "string",
            "example": "2026-04-27T23:59:59Z"
          },
          "actionReportDeadline": {
            "type": "string",
            "example": "2026-04-29T23:59:59Z"
          }
        },
        "required": [
          "cycleName",
          "startTimestamp",
          "endTimestamp"
        ]
      },
      "UpdateFeedbackCycleDto": {
        "type": "object",
        "properties": {
          "cycleName": {
            "type": "string"
          },
          "type": {
            "type": "string"
          },
          "startTimestamp": {
            "type": "string"
          },
          "endTimestamp": {
            "type": "string"
          },
          "reflectionDeadline": {
            "type": "string"
          },
          "actionReportDeadline": {
            "type": "string"
          }
        }
      },
      "UpdateCycleStatusDto": {
        "type": "object",
        "properties": {
          "status": {
            "type": "string",
            "example": "active",
            "enum": [
              "active",
              "closed",
              "preparation"
            ]
          },
          "phase": {
            "type": "string",
            "example": "STUDENT_FEEDBACK",
            "description": "Phase of the cycle"
          }
        },
        "required": [
          "status"
        ]
      },
      "BulkImportCyclesDto": {
        "type": "object",
        "properties": {
          "cycles": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/CreateFeedbackCycleDto"
            }
          }
        },
        "required": [
          "cycles"
        ]
      },
...
    }
  }
}